requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==2.3.7
opencv-python==4.8.1.78
numpy==1.24.3
pandas==1.5.3
face-recognition==1.3.0
Pillow==10.0.1
insightface==0.7.3
onnxruntime==1.16.3
scikit-learn==1.3.2
python-dotenv==1.0.0
bcrypt==4.1.2
